+++ skydrop/src/fc/logger/logger.cpp
@@ -11,7 +11,7 @@ FIL log_file;
 uint32_t logger_next = 0;
 uint8_t logger_flight_number;
 uint32_t logger_flight_day;
-int16_t logger_last_temp;
+int8_t logger_last_temp;
 
 void logger_init()
 {
